fix several type incompatibilities in postgres + sqlite interop cases

+++ backend/src/db/builds.rs
@@ -13,8 +13,8 @@ pub struct Model {
     pub pkg_id: i32,
     pub output: Option<String>,
     pub status: Option<i32>,
-    pub start_time: Option<u32>,
-    pub end_time: Option<u32>,
+    pub start_time: Option<i64>,
+    pub end_time: Option<i64>,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
